% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afcov.r
\name{afcov}
\alias{afcov}
\title{to return subnational covid data by country and optionally date}
\usage{
afcov(country, language = "en", dates = "all")
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{language}{'which admin level to return 'en' or 'fr' for country name}

\item{dates}{either 'all', 'last' or c(start,end) in format "2021-01-31"}
}
\value{
\code{ggplot}
}
\description{
*in progress
}
