% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{dfhera}
\alias{dfhera}
\title{subnational covid data from HERA}
\format{
\code{dataframe}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 66983 rows and 17 columns.
}
\source{
\url{https://data.humdata.org/organization/hera-humanitarian-emergency-response-africa}
}
\usage{
dfhera
}
\description{
subnational covid data from HERA
}
\section{Slots}{

\describe{
\item{\code{data}}{subnational daily covid statistics}
}}

\keyword{datasets}
