% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_subnat_to_map.r
\name{join_subnat_to_map}
\alias{join_subnat_to_map}
\title{join subnational data to map of admin boundaries}
\usage{
join_subnat_to_map(country = "senegal")
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}
}
\value{
dataframe
}
\description{
in progress
}
