% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afcov_heatmap.r
\name{afcov_heatmap}
\alias{afcov_heatmap}
\title{plot heatmap of subnational covid data}
\usage{
afcov_heatmap(
  country,
  attribute = "CONTAMINES",
  language = "en",
  legend.position = "top"
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{attribute}{attribute to plot, from dfhera}

\item{language}{'which admin level to return'en' or 'fr' for country name}

\item{legend.position}{default "top", options "left","right","bottom"}
}
\value{
\code{ggplot}
}
\description{
*in progress
to show cases or deaths
}
