% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afcov_heatmap.r
\name{afcov_heatmap}
\alias{afcov_heatmap}
\title{plot heatmap of subnational covid data}
\usage{
afcov_heatmap(
  country,
  attribute = "CONTAMINES",
  language = "en",
  legend.position = "top",
  date_legend = c("2020-04", "2020-05", "2020-06", "2020-07", "2020-08", "2020-09",
    "2020-10", "2020-11", "2020-12", "2021-01", "2021-02", "2021-03")
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{attribute}{attribute to plot, from dfhera}

\item{language}{'which admin level to return'en' or 'fr' for country name}

\item{legend.position}{default "top", options "left","right","bottom"}

\item{date_legend}{axis breaks to appear in plots}
}
\value{
\code{ggplot}
}
\description{
*in progress
to show cases or deaths
}
