% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afcov_map.r
\name{afcov_map}
\alias{afcov_map}
\title{plot map of subnational covid data by country}
\usage{
afcov_map(
  country,
  attribute = "CONTAMINES",
  dates = "last",
  timeinterval = NULL,
  language = "en",
  legend.position = "top",
  palette = "YlGnBu"
)
}
\arguments{
\item{country}{a character vector of country names or iso3c character codes.}

\item{attribute}{attribute to plot, from dfhera}

\item{dates}{either 'all', 'last' or c(start,end) in format "2021-01-31"}

\item{timeinterval}{optional interval to sum over, options e.g. "5 days" "week" "month"}

\item{language}{'which admin level to return 'en' or 'fr' for country name}

\item{legend.position}{default "top", options "left","right","bottom"}

\item{palette}{RColorBrewer palette options default="YlGnBu"}
}
\value{
\code{ggplot}
}
\description{
*in progress
to show cases or deaths
}
\examples{
afcov_map('Mali')

#a 6 daily facetted map
afcov_map('senegal', dates=c("2021-01-01","2021-01-06"))

}
