% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_all_subnat_to_map.r
\name{join_all_subnat_to_map}
\alias{join_all_subnat_to_map}
\title{join geoboundaries admin names onto the hera data to allow joining to a map later}
\usage{
join_all_subnat_to_map()
}
\value{
dataframe
}
\description{
*in progress
}
