% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.r
\name{update_data}
\alias{update_data}
\title{update package data from HERA via HDX}
\usage{
update_data()
}
\value{
dataframe
}
\description{
*in progress
}
\details{
will try to download data from HDX and use this to update locally the data stored in the package
}
